+++ core/src/main/java/org/bptlab/cepta/operators/SumOfDelayAtStationFunction.java
@@ -2,15 +2,15 @@ package org.bptlab.cepta.operators;
 
 import java.lang.Object;
 import java.util.HashMap;
-import java.lang.reflect.*;
 import org.apache.flink.api.common.functions.MapFunction;
 import org.apache.flink.api.common.functions.FilterFunction;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.api.java.tuple.Tuple2;
 
+import org.bptlab.cepta.models.internal.notifications.notification.NotificationOuterClass;
+
 import org.apache.flink.util.Collector;
 
-import org.bptlab.cepta.models.internal.delay.DelayOuterClass;
 import org.bptlab.cepta.utils.triggers.CustomCountTrigger;
 import org.apache.flink.streaming.api.functions.windowing.ProcessWindowFunction;
 import org.apache.flink.streaming.api.windowing.windows.GlobalWindow;
@@ -18,19 +18,19 @@ import org.apache.flink.streaming.api.windowing.assigners.GlobalWindows;
 
 import org.apache.flink.api.java.functions.KeySelector;
 
-public class SumOfDelayAtStationFunction<T extends Object> {
+public class SumOfDelayAtStationFunction {
 
     /* This function expects a Stream of TrainDelayNotification events and a window size
     and sums up all delay based on the location Id's in the given window size.
     The window is a fixed event number window.
     It will return a Stream of Tuple2 with the location Id and the sum of delay.
     */
-    public DataStream<Tuple2<Integer, Long>> SumOfDelayAtStation(DataStream<T> inputStream, int windowSize, String stationAttributName) {
-        DataStream<Tuple2<Integer, Long>> resultStream = inputStream
+    public DataStream<Tuple2<Long, Double>> SumOfDelayAtStation(DataStream<NotificationOuterClass.DelayNotification> inputStream, int windowSize) {
+        DataStream<Tuple2<Long, Double>> resultStream = inputStream
                 .keyBy(
-                        new KeySelector<T, Integer>(){
-                            Integer key = Integer.valueOf(0);
-                            public Integer getKey(T event){
+                        new KeySelector<NotificationOuterClass.DelayNotification, Integer>(){
+                            Integer key = 0;
+                            public Integer getKey(NotificationOuterClass.DelayNotification event){
                                 Integer returnKey = key/windowSize;
                                 key++;
                                 return returnKey;
@@ -40,38 +40,33 @@ public class SumOfDelayAtStationFunction<T extends Object> {
                 .window(GlobalWindows.create()).trigger(
                         CustomCountTrigger.of(windowSize)
                 ).process(
-                        sumOfDelayAtStationWindowProcessFunction(stationAttributName)
+                        sumOfDelayAtStationWindowProcessFunction()
                 );
         return resultStream;
     };
 
-    public ProcessWindowFunction<T, Tuple2<Integer, Long>, Integer, GlobalWindow> sumOfDelayAtStationWindowProcessFunction(String stationAttributName) {
-        return new ProcessWindowFunction<T, Tuple2<Integer, Long>, Integer, GlobalWindow>() {
+    public static ProcessWindowFunction<NotificationOuterClass.DelayNotification, Tuple2<Long, Double>, Integer, GlobalWindow> sumOfDelayAtStationWindowProcessFunction() {
+        return new ProcessWindowFunction<NotificationOuterClass.DelayNotification, Tuple2<Long, Double>, Integer, GlobalWindow>() {
             @Override
-            public void process(Integer key, Context context, Iterable<T> input, Collector<Tuple2<Integer, Long>> out) throws Exception {
-                HashMap<Integer, Long> sums = new HashMap<Integer, Long>();
-                for (T in: input) {
-                    // This code is very very very bad
-                    System.out.println(in);
-                    Class c = in.getClass();
-                    String methodName = "get" + stationAttributName;
-                    Method method = c.getDeclaredMethod(methodName);
-                    Integer stationId = Integer.valueOf(method.invoke(in).toString());
-                    method = c.getDeclaredMethod("getDelay");
-                    DelayOuterClass.Delay delay = (DelayOuterClass.Delay) method.invoke(in);
-
-                    if (!sums.containsKey(stationId)) {
-                        sums.put(stationId, delay.getDelta().getSeconds());
+            public void process(Integer key, Context context, Iterable<NotificationOuterClass.DelayNotification> input, Collector<Tuple2<Long, Double>> out) throws Exception {
+                HashMap<Long, Double> sums = new HashMap<Long, Double>();
+                for (NotificationOuterClass.DelayNotification in: input) {
+                    Long trainId = Long.valueOf(in.getTransportId().getId());
+                    Long locationId = Long.valueOf(in.getStationId().getId());
+                    Double delay = (double) in.getDelay().getDelta().getSeconds();
+                    if (!sums.containsKey(locationId)) {
+                        sums.put(locationId, delay);
                     } else {
-                        long tmp;
-                        tmp = sums.get(stationId);
-                        sums.replace(stationId, (tmp + delay.getDelta().getSeconds()));
+                        double tmp;
+                        tmp = sums.get(locationId);
+                        sums.replace(locationId, (tmp + delay));
                     }
                 }
 
-                for (Integer station: sums.keySet()) {
-                    long delay = sums.get(station);
-                    out.collect(new Tuple2<Integer, Long>(station, delay) );
+                for (Long location: sums.keySet()) {
+
+                    Double delay = sums.get(location);
+                    out.collect(new Tuple2<Long, Double>(location, delay) );
                 }
             }
         };