+++ core/src/test/java/org/bptlab/cepta/sum-of-delay-at-station/SumOfDelayAtStationTests.java
@@ -3,8 +3,6 @@ package org.bptlab.cepta;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.io.IOException;
-
-import org.bptlab.cepta.models.internal.notifications.notification.NotificationOuterClass;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.Ignore;
@@ -16,24 +14,20 @@ import org.apache.flink.api.java.tuple.Tuple2;
 import org.bptlab.cepta.operators.SumOfDelayAtStationFunction;
 
 import org.bptlab.cepta.providers.TrainDelayNotificationDataProvider;
-import org.bptlab.cepta.providers.LiveTrainDataProvider;
 
-import org.bptlab.cepta.models.events.train.LiveTrainDataOuterClass.LiveTrainData;
-
-import javax.management.Notification;
+import org.bptlab.cepta.models.internal.notifications.notification.NotificationOuterClass;
 
 public class SumOfDelayAtStationTests {
     @Test
-    public void TestSumOfDelaysAtStationWithTrainDelayNotification() throws IOException {
-        if (true) return;
+    public void TestSumOfDelaysAtStation() throws IOException {
 
         boolean pass = true;
-        Integer expectedStation1 = 1;
-        Integer expectedStation2 = 2;
-        Long expectedDelayAtStation1 = 25L;
-        Long expectedDelayAtStation2 = 13L;
+        Long expectedStation1 = 1L;
+        Long expectedStation2 = 2L;
+        Double expectedDelayAtStation1 = 25.0;
+        Double expectedDelayAtStation2 = 13.0;
 
-        SumOfDelayAtStationFunction sumOfDelayAtStationFunction = new SumOfDelayAtStationFunction<NotificationOuterClass.DelayNotification>();
+        SumOfDelayAtStationFunction sumOfDelayAtStationFunction = new SumOfDelayAtStationFunction();
         // the provider provides four TrainDelayNotification elements
         // element 1 has stationId 1, trainId 1, delay 10
         // element 2 has stationId 2, trainId 2, delay 5
@@ -41,25 +35,21 @@ public class SumOfDelayAtStationTests {
         // element 4 has stationId 2, trainId 1, delay 8
         DataStream<NotificationOuterClass.DelayNotification> delayNotificationStream = TrainDelayNotificationDataProvider.TrainDelayNotificationDataStream();
 
-        DataStream<Tuple2<Integer, Long>> stationAndDelayStream = sumOfDelayAtStationFunction.SumOfDelayAtStation(delayNotificationStream, 4, "StationId");
-        ArrayList<Tuple2<Integer, Long>> stationAndDelayArray = new ArrayList<>();
-        Iterator<Tuple2<Integer, Long>> iterator = DataStreamUtils.collect(stationAndDelayStream);
+        DataStream<Tuple2<Long, Double>> locationAndDelayStream = sumOfDelayAtStationFunction.SumOfDelayAtStation(delayNotificationStream, 4);
+        ArrayList<Tuple2<Long, Double>> locationAndDelayArray = new ArrayList<>();
+        Iterator<Tuple2<Long, Double>> iterator = DataStreamUtils.collect(locationAndDelayStream);
         while(iterator.hasNext()){
-            Tuple2<Integer, Long> tuple = iterator.next();
-            stationAndDelayArray.add(tuple);
+            Tuple2<Long, Double> tuple = iterator.next();
+            locationAndDelayArray.add(tuple);
         }
-        // check if any tuple is present
-        if (stationAndDelayArray.size() == 0) {
-            pass = false;
-        }
-        for (Tuple2<Integer, Long> tuple : stationAndDelayArray) {
-            // check if first station matches expected delay
+
+        for (Tuple2<Long, Double> tuple : locationAndDelayArray) {
             if (tuple.f0.equals(expectedStation1)) {
                 if (!tuple.f1.equals(expectedDelayAtStation1)) {
                     pass = false;
                 }
             }
-            // check if second station matches expected delay
+
             if (tuple.f0.equals(expectedStation2)) {
                 if (!tuple.f1.equals(expectedDelayAtStation2)) {
                     pass = false;
@@ -67,72 +57,6 @@ public class SumOfDelayAtStationTests {
             }
         }
 
-        Assert.assertTrue(pass);
-    }
-
-    @Test
-    public void TestSumOfDelaysAtStationWithLiveTrainData() throws IOException {
-
-        boolean pass = true;
-        Integer expectedStation1 = 1;
-        Long expectedDelayAtStation1 = 3L;
-
-        SumOfDelayAtStationFunction sumOfDelayAtStationFunction = new SumOfDelayAtStationFunction<LiveTrainData>();
-        DataStream<LiveTrainData> delayNotificationStream = LiveTrainDataProvider.liveTrainDatStreamWithDuplicates();
-
-        DataStream<Tuple2<Integer, Long>> stationAndDelayStream = sumOfDelayAtStationFunction.SumOfDelayAtStation(delayNotificationStream, 3, "StationId");
-        ArrayList<Tuple2<Integer, Long>> stationAndDelayArray = new ArrayList<>();
-        Iterator<Tuple2<Integer, Long>> iterator = DataStreamUtils.collect(stationAndDelayStream);
-        while(iterator.hasNext()){
-            Tuple2<Integer, Long> tuple = iterator.next();
-            stationAndDelayArray.add(tuple);
-        }
-
-        System.out.println(stationAndDelayArray);
-
-        // check if any tuple is present
-        if (stationAndDelayArray.size() == 0) {
-            pass = false;
-        }
-
-        for (Tuple2<Integer, Long> tuple : stationAndDelayArray) {
-            // check if first station matches expected delay
-            if (tuple.f0.equals(expectedStation1)) {
-                if (!tuple.f1.equals(expectedDelayAtStation1)) {
-                    pass = false;
-                }
-            }
-        }
-        Assert.assertTrue(pass);
-    }
-
-    @Test
-    public void TestSumOfDelaysAtStationRightNumberOfTuples() throws IOException {
-        if (true) return;
-        boolean pass = true;
-        SumOfDelayAtStationFunction sumOfDelayAtStationFunction = new SumOfDelayAtStationFunction<NotificationOuterClass.DelayNotification>();
-        // the provider provides four TrainDelayNotification elements
-        // element 1 has stationId 1, trainId 1, delay 10
-        // element 2 has stationId 2, trainId 2, delay 5
-        // element 3 has stationId 1, trainId 2, delay 15
-        // element 4 has stationId 2, trainId 1, delay 8
-        DataStream<NotificationOuterClass.DelayNotification> delayNotificationStream = TrainDelayNotificationDataProvider.TrainDelayNotificationDataStream();
-
-        DataStream<Tuple2<Integer, Long>> stationAndDelayStream = sumOfDelayAtStationFunction.SumOfDelayAtStation(delayNotificationStream, 4, "StationId");
-        ArrayList<Tuple2<Integer, Long>> locationAndDelayArray = new ArrayList<>();
-        Iterator<Tuple2<Integer, Long>> iterator = DataStreamUtils.collect(stationAndDelayStream);
-        while(iterator.hasNext()){
-            Tuple2<Integer, Long> tuple = iterator.next();
-            locationAndDelayArray.add(tuple);
-        }
-        // check if any tuple is present
-        if (locationAndDelayArray.size() == 0) {
-            pass = false;
-        }
-        // check if there are only 2 Tuples present
-        if (locationAndDelayArray.size() != 2) {
-            pass = false;
-        }
         Assert.assertTrue(pass);
     }
 }