+++ core/src/test/java/org/bptlab/cepta/weather-tests/WeatherLiveTrainJoinTests.java
@@ -2,6 +2,9 @@ package org.bptlab.cepta;
 
 import java.io.IOException;
 import java.util.Iterator;
+import java.util.stream.Stream;
+
+import org.bptlab.cepta.utils.functions.StreamUtils;
 import org.javatuples.Pair;
 import org.junit.Assert;
 import org.junit.Test;
@@ -31,13 +34,9 @@ public class WeatherLiveTrainJoinTests {
     DataStream<NotificationOuterClass.Notification> trainDelayNotificationDataStream =
         WeatherLiveTrainJoinFunction.delayFromWeather(correlatedWeatherStream, liveTrainStream);
 
-    Iterator<NotificationOuterClass.Notification> iterator = DataStreamUtils.collect(trainDelayNotificationDataStream);
-    int count = 0;
-    while(iterator.hasNext()){
-      iterator.next();
-      count++;
-    }
-    Assert.assertEquals(count, 1);
+    int count = StreamUtils.countOfEventsInStream(trainDelayNotificationDataStream);
+
+    Assert.assertEquals(1, count);
   }
 
   @Test
@@ -49,13 +48,9 @@ public class WeatherLiveTrainJoinTests {
     DataStream<NotificationOuterClass.Notification> trainDelayNotificationDataStream =
         WeatherLiveTrainJoinFunction.delayFromWeather(correlatedWeatherStream, liveTrainStream);
 
-    Iterator<NotificationOuterClass.Notification> iterator = DataStreamUtils.collect(trainDelayNotificationDataStream);
-    int count = 0;
-    while(iterator.hasNext()){
-      iterator.next();
-      count++;
-    }
-    Assert.assertEquals(count, 4);
+    int count = StreamUtils.countOfEventsInStream(trainDelayNotificationDataStream);
+
+    Assert.assertEquals(4, count);
   }
 
   @Test
@@ -67,12 +62,7 @@ public class WeatherLiveTrainJoinTests {
     DataStream<NotificationOuterClass.Notification> trainDelayNotificationDataStream =
         WeatherLiveTrainJoinFunction.delayFromWeather(correlatedWeatherStream, liveTrainStream);
 
-    Iterator<NotificationOuterClass.Notification> iterator = DataStreamUtils.collect(trainDelayNotificationDataStream);
-    int count = 0;
-    while(iterator.hasNext()){
-      iterator.next();
-      count++;
-    }
+    int count = StreamUtils.countOfEventsInStream(trainDelayNotificationDataStream);
     Assert.assertEquals(count, 0);
   }
 }