+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 ### [0.4.0] - 2020-05-02
 ##### Added
 - WIP
+
+- [#248](https://github.com/bptlab/cepta/pull/248)
+  - Refactors Delay detection based on Live and Planned Station arrival Data including tests
 - [#259](https://github.com/bptlab/cepta/pull/259)
   - Integration of the `usermgmt` service with the frontend via module
   - Display information about the current user
   - Edit and remove own user functionality
   - New *AddUser* view
+- [#246] (https://github.com/bptlab/cepta/pull/246)
+  - operator that gets delay from live train and shifts it by creating delay notifications for subsequent stations
 - [#240](https://github.com/bptlab/cepta/pull/240)
   - Subscription to correct Kafka queue
   - Connection to UserDB
@@ -20,6 +25,8 @@ All notable changes to this project will be documented in this file.
 
 ##### Changed
 - WIP
+- [#248](https://github.com/bptlab/cepta/pull/248)
+  - changes TrainDelayNotification location_id to station_id to match conventions of other events
 - [#220](https://github.com/bptlab/cepta/pull/220)
   - generic internal high level models to be used in ESP pipelines
 - [#213](https://github.com/bptlab/cepta/pull/213)